🔨refactor: update CLI paths and imports for `mindnum`
- refactor CLI main path to include `mindnum` directory
- update import paths in CLI commands and formatter
- modify `README.md` to reflect new installation paths
- ensure all tests and documentation reflect the new structure
