+++ .goreleaser.yaml
@@ -7,7 +7,7 @@ before:
     - go mod tidy
 
 builds:
-  - main: ./app/presentation/cli/main.go
+  - main: ./app/presentation/cli/mindnum/main.go
     env:
       - CGO_ENABLED=0
     goos:
