+++ README.md
@@ -53,20 +53,8 @@ Flags:
 
 ### 🐭 Using go
 
-#### If you have [gostall](https://github.com/davidmdm/gostall), execute below.
-
 ```
-GOBIN=$GOPATH/bin gostall github.com/yanosea/mindnum/app/presentation/cli@latest mindnum
-```
-
-#### Or, execute below.
-
-```
-# This will install mindnum as the name cli to $GOPATH/bin, so please change the name.
-go install github.com/yanosea/mindnum/app/presentation/cli@latest 
-
-# Change the name to mindnum.
-mv $GOPATH/bin/cli $GOPATH/bin/mindnum
+go install github.com/yanosea/mindnum/app/presentation/cli/mindnum@latest 
 ```
 
 ### 🍺 Using homebrew
@@ -86,20 +74,8 @@ Go to the [Releases](https://github.com/yanosea/mindnum/releases) and download t
 
 Reinstall `mindnum`!
 
-#### If you have [gostall](https://github.com/davidmdm/gostall), execute below.
-
 ```
-GOBIN=$GOPATH/bin gostall github.com/yanosea/mindnum/app/presentation/cli@latest mindnum
-```
-
-#### Or, execute below.
-
-```
-# This will install mindnum as the name cli to $GOPATH/bin, so please change the name.
-go install github.com/yanosea/mindnum/app/presentation/cli@latest 
-
-# Change the name to mindnum.
-mv $GOPATH/bin/cli $GOPATH/bin/mindnum
+go install github.com/yanosea/mindnum/app/presentation/cli/mindnum@latest 
 ```
 
 ### 🍺 Using homebrew
