+++ app/presentation/cli/mindnum/command/command_test.go
