+++ app/presentation/cli/mindnum/command/mindnum/get_test.go
