+++ app/presentation/cli/mindnum/command/mindnum/godoc.go
