+++ app/presentation/cli/mindnum/command/mindnum/version_test.go
