+++ app/presentation/cli/mindnum/command/root_test.go
