+++ app/presentation/cli/mindnum/formatter/color.go
