+++ app/presentation/cli/mindnum/formatter/formatter_test.go
