+++ app/presentation/cli/mindnum/formatter/text.go
