+++ app/presentation/cli/mindnum/formatter/text_test.go
