+++ app/presentation/cli/mindnum/godoc.go
