+++ app/presentation/cli/mindnum/presenter/mindnum_presenter_test.go
